+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @puzzlehq/sdk-core
 
+## 0.5.6
+
+### Patch Changes
+
+- improve state management for event creation
+
 ## 0.5.5
 
 ### Patch Changes